perf(documents): lazy-load react-pdf viewer on demand (#1639)
Split PDF rendering into a dynamically imported PdfDocumentViewer so
react-pdf and its worker load only when a PDF is opened, keeping the
library out of shared first-load chunks.
