+++ frontend/components/documents/PdfDocumentViewer.tsx
@@ -1,0 +1,131 @@
+'use client';
+
+import React, { useState } from 'react';
+import { Document, Page, pdfjs } from 'react-pdf';
+import { ChevronLeft, ChevronRight, ZoomIn, ZoomOut } from 'lucide-react';
+
+// Load the PDF.js worker on demand with the dynamically imported module.
+pdfjs.GlobalWorkerOptions.workerSrc = `//unpkg.com/pdfjs-dist@${pdfjs.version}/build/pdf.worker.min.mjs`;
+
+interface PdfDocumentViewerProps {
+  url: string;
+  name: string;
+}
+
+export default function PdfDocumentViewer({ url, name }: PdfDocumentViewerProps) {
+  const [numPages, setNumPages] = useState<number>(0);
+  const [pageNumber, setPageNumber] = useState<number>(1);
+  const [scale, setScale] = useState<number>(1.0);
+  const [loading, setLoading] = useState<boolean>(true);
+  const [error, setError] = useState<string | null>(null);
+
+  const onDocumentLoadSuccess = ({ numPages: totalPages }: { numPages: number }) => {
+    setNumPages(totalPages);
+    setLoading(false);
+    setError(null);
+  };
+
+  const onDocumentLoadError = (loadError: Error) => {
+    setLoading(false);
+    setError('Failed to load document. Please try again.');
+    console.error('PDF load error:', loadError);
+  };
+
+  const goToPrevPage = () => {
+    setPageNumber((prev) => Math.max(prev - 1, 1));
+  };
+
+  const goToNextPage = () => {
+    setPageNumber((prev) => Math.min(prev + 1, numPages));
+  };
+
+  const zoomIn = () => {
+    setScale((prev) => Math.min(prev + 0.2, 3.0));
+  };
+
+  const zoomOut = () => {
+    setScale((prev) => Math.max(prev - 0.2, 0.5));
+  };
+
+  return (
+    <div className="space-y-4">
+      {!loading && !error && (
+        <div className="flex items-center justify-between bg-neutral-100 dark:bg-neutral-800 rounded-xl p-3">
+          <div className="flex items-center gap-2">
+            <button
+              onClick={goToPrevPage}
+              disabled={pageNumber <= 1}
+              className="p-2 rounded-lg bg-white dark:bg-neutral-700 hover:bg-neutral-200 dark:hover:bg-neutral-600 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
+            >
+              <ChevronLeft size={18} />
+            </button>
+            <span className="text-sm font-medium text-neutral-900 dark:text-neutral-100 min-w-[100px] text-center">
+              Page {pageNumber} of {numPages}
+            </span>
+            <button
+              onClick={goToNextPage}
+              disabled={pageNumber >= numPages}
+              className="p-2 rounded-lg bg-white dark:bg-neutral-700 hover:bg-neutral-200 dark:hover:bg-neutral-600 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
+            >
+              <ChevronRight size={18} />
+            </button>
+          </div>
+
+          <div className="flex items-center gap-2">
+            <button
+              onClick={zoomOut}
+              disabled={scale <= 0.5}
+              className="p-2 rounded-lg bg-white dark:bg-neutral-700 hover:bg-neutral-200 dark:hover:bg-neutral-600 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
+            >
+              <ZoomOut size={18} />
+            </button>
+            <span className="text-sm font-medium text-neutral-900 dark:text-neutral-100 min-w-[60px] text-center">
+              {Math.round(scale * 100)}%
+            </span>
+            <button
+              onClick={zoomIn}
+              disabled={scale >= 3.0}
+              className="p-2 rounded-lg bg-white dark:bg-neutral-700 hover:bg-neutral-200 dark:hover:bg-neutral-600 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
+            >
+              <ZoomIn size={18} />
+            </button>
+          </div>
+        </div>
+      )}
+
+      <div className="flex items-center justify-center bg-neutral-100 dark:bg-neutral-900 rounded-xl overflow-auto min-h-[400px] max-h-[600px]">
+        {loading && (
+          <div className="text-center py-12">
+            <div className="inline-block w-8 h-8 border-4 border-brand-blue border-t-transparent rounded-full animate-spin mb-4"></div>
+            <p className="text-sm text-neutral-600 dark:text-neutral-400">
+              Loading document...
+            </p>
+          </div>
+        )}
+
+        {error && (
+          <div className="text-center py-12">
+            <p className="text-sm text-red-600 dark:text-red-400">{error}</p>
+          </div>
+        )}
+
+        {!error && (
+          <Document
+            file={url}
+            onLoadSuccess={onDocumentLoadSuccess}
+            onLoadError={onDocumentLoadError}
+            loading=""
+            error=""
+          >
+            <Page
+              pageNumber={pageNumber}
+              scale={scale}
+              renderTextLayer={true}
+              renderAnnotationLayer={true}
+            />
+          </Document>
+        )}
+      </div>
+    </div>
+  );
+}
